Added signals class
We can't 'self.emit' from SimpleDBusNotifications. We have to use this Signals class to pass filtered notification signals. The user could just as well connect to 'g-signals' but then they'd have to filter out the signals that don't pertain to their notification and filter the signal it self. This is just much easier for them.